% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_lm_assumptions.R
\name{check_lm_assumptions}
\alias{check_lm_assumptions}
\title{Check Linear Regression Model Assumptions}
\usage{
check_lm_assumptions(model, m = NULL)
}
\arguments{
\item{model}{Object of class 'lm'. The fitted linear regression model.}

\item{m}{Number of digits to round the results to. Default is NULL.}
}
\value{
A list containing the results of assumption checks.
}
\description{
This function verifies assumptions for various linear regression models, including OLS, WLS, censored, and truncated models.
}
