% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegAssure-package.R
\docType{package}
\name{RegAssure-package}
\alias{RegAssure}
\alias{RegAssure-package}
\title{RegAssure: Streamlined Integration of Regression Assumption}
\description{
It streamlines the evaluation of regression model assumptions, enhancing result reliability. With integrated tools for assessing key aspects like linearity, homoscedasticity, and more. It's a valuable asset for researchers and analysts working with regression models.
}
\author{
\strong{Maintainer}: Nicolás Rubio García \email{nrubiogar@gmail.com} (\href{https://orcid.org/0000-0003-3502-0784}{ORCID})

}
\keyword{internal}
